Corrected Doubles to Ints in parseXML.java

+++ src/io/parseXML.java
@@ -274,11 +274,11 @@ public class parseXML {
                 	   Assignment currentAssignment = currentCategory.getAssignment(currentCategory.getAssignmentIndex(assignmentName));
                 	   if (studentIndex < totalStudents) {
                 		   String studentPseudoName = course.getStudent(studentIndex).getPseudoName();
-                		   currentAssignment.setGrade(studentPseudoName, Double.parseDouble(new String(ch, start, length)));
+                		   currentAssignment.setGrade(studentPseudoName, Integer.parseInt(new String(ch, start, length)));
                 	   }
                 	   else {
                 		   String ghostPseudoName = course.getGhostStudent(studentIndex - totalStudents).getPseudoName();
-                		   currentAssignment.setGrade(ghostPseudoName, Double.parseDouble(new String(ch, start, length)));
+                		   currentAssignment.setGrade(ghostPseudoName, Integer.parseInt(new String(ch, start, length)));
                 	   }
                 	   isGrade = false;
                 	   studentIndex ++;
